pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "karsheft-flock"
version = "0.1.0"
description = "Reproducible Debian package management with cryptographic verification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "tomli>=2.0; python_version < '3.11'",
    "tomli-w>=1.0",
    "requests>=2.28",
    "python-gnupg>=0.5",
]

[project.scripts]
flock = "flock.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
